+++ src/libexpr-c/nix_api_value.cc
@@ -125,7 +125,7 @@ PrimOp * nix_alloc_primop(
     try {
         using namespace std::placeholders;
         auto p = new
-#if HAVE_BOEHMGC
+#if NIX_USE_BOEHMGC
             (GC)
 #endif
                 nix::PrimOp{
@@ -497,7 +497,7 @@ ListBuilder * nix_make_list_builder(nix_c_context * context, EvalState * state, 
     try {
         auto builder = state->state.buildList(capacity);
         return new
-#if HAVE_BOEHMGC
+#if NIX_USE_BOEHMGC
             (NoGC)
 #endif
                 ListBuilder{std::move(builder)};
@@ -519,7 +519,7 @@ nix_list_builder_insert(nix_c_context * context, ListBuilder * list_builder, uns
 
 void nix_list_builder_free(ListBuilder * list_builder)
 {
-#if HAVE_BOEHMGC
+#if NIX_USE_BOEHMGC
     GC_FREE(list_builder);
 #else
     delete list_builder;
@@ -578,7 +578,7 @@ BindingsBuilder * nix_make_bindings_builder(nix_c_context * context, EvalState *
     try {
         auto bb = state->state.buildBindings(capacity);
         return new
-#if HAVE_BOEHMGC
+#if NIX_USE_BOEHMGC
             (NoGC)
 #endif
                 BindingsBuilder{std::move(bb)};
@@ -600,7 +600,7 @@ nix_err nix_bindings_builder_insert(nix_c_context * context, BindingsBuilder * b
 
 void nix_bindings_builder_free(BindingsBuilder * bb)
 {
-#if HAVE_BOEHMGC
+#if NIX_USE_BOEHMGC
     GC_FREE((nix::BindingsBuilder *) bb);
 #else
     delete (nix::BindingsBuilder *) bb;
