+++ src/libexpr/meson.build
@@ -14,7 +14,8 @@ cxx = meson.get_compiler('cpp')
 
 subdir('nix-meson-build-support/deps-lists')
 
-configdata = configuration_data()
+configdata_pub = configuration_data()
+configdata_priv = configuration_data()
 
 deps_private_maybe_subproject = [
 ]
@@ -25,6 +26,16 @@ deps_public_maybe_subproject = [
 ]
 subdir('nix-meson-build-support/subprojects')
 subdir('nix-meson-build-support/big-objs')
+
+# Check for each of these functions, and create a define like `#define HAVE_LCHOWN 1`.
+check_funcs = [
+  'sysconf',
+]
+foreach funcspec : check_funcs
+  define_name = 'HAVE_' + funcspec.underscorify().to_upper()
+  define_value = cxx.has_function(funcspec).to_int()
+  configdata_priv.set(define_name, define_value)
+endforeach
 
 boost = dependency(
   'boost',
@@ -47,11 +58,13 @@ if bdw_gc.found()
   ]
     define_name = 'HAVE_' + funcspec.underscorify().to_upper()
     define_value = cxx.has_function(funcspec).to_int()
-    configdata.set(define_name, define_value)
+    configdata_priv.set(define_name, define_value)
   endforeach
-  configdata.set('GC_THREADS', 1)
+  # Affects ABI, because it changes what bdw_gc itself does!
+  configdata_pub.set('GC_THREADS', 1)
 endif
-configdata.set('HAVE_BOEHMGC', bdw_gc.found().to_int())
+# Used in public header. Affects ABI!
+configdata_pub.set('NIX_USE_BOEHMGC', bdw_gc.found().to_int())
 
 toml11 = dependency(
   'toml11',
@@ -61,14 +74,9 @@ toml11 = dependency(
 )
 deps_other += toml11
 
-add_project_arguments(
-  # TODO(Qyriad): Yes this is how the autoconf+Make system did it.
-  # It would be nice for our headers to be idempotent instead.
-  '-include', 'nix/config-util.hh',
-  '-include', 'nix/config-store.hh',
-  # '-include', 'nix_api_fetchers_config.h',
-  '-include', 'nix/config-expr.hh',
-  language : 'cpp',
+config_priv_h = configure_file(
+  configuration : configdata_priv,
+  output : 'expr-config-private.hh',
 )
 
 subdir('nix-meson-build-support/common')
@@ -158,6 +166,7 @@ subdir('nix-meson-build-support/windows-version')
 this_library = library(
   'nixexpr',
   sources,
+  config_priv_h,
   parser_tab,
   lexer_tab,
   generated_headers,
