+++ src/libmain/include/nix/meson.build
@@ -2,12 +2,7 @@
 
 include_dirs = [include_directories('..')]
 
-config_h = configure_file(
-  configuration : configdata,
-  output : 'config-main.hh',
-)
-
-headers = [config_h] + files(
+headers = files(
   'common-args.hh',
   'loggers.hh',
   'plugin.hh',
