+++ src/libmain/shared.cc
@@ -25,6 +25,8 @@
 #include "nix/exit.hh"
 #include "nix/strings.hh"
 
+#include "main-config-private.hh"
+
 namespace nix {
 
 char * * savedArgv;
@@ -297,7 +299,7 @@ void printVersion(const std::string & programName)
     std::cout << fmt("%1% (Nix) %2%", programName, nixVersion) << std::endl;
     if (verbosity > lvlInfo) {
         Strings cfg;
-#if HAVE_BOEHMGC
+#if NIX_USE_BOEHMGC
         cfg.push_back("gc");
 #endif
         cfg.push_back("signed-caches");
