+++ src/libstore-c/meson.build
@@ -23,17 +23,6 @@ deps_public_maybe_subproject = [
 ]
 subdir('nix-meson-build-support/subprojects')
 
-add_project_arguments(
-  # TODO(Qyriad): Yes this is how the autoconf+Make system did it.
-  # It would be nice for our headers to be idempotent instead.
-
-  # From C++ libraries, only for internals
-  '-include', 'nix/config-util.hh',
-  '-include', 'nix/config-store.hh',
-
-  language : 'cpp',
-)
-
 subdir('nix-meson-build-support/common')
 
 sources = files(
