+++ src/libutil/compute-levels.cc
@@ -1,8 +1,10 @@
 #include "nix/types.hh"
 
+#include "util-config-private.hh"
+
 #if HAVE_LIBCPUID
-#include <libcpuid/libcpuid.h>
-#include <map>
+#  include <libcpuid/libcpuid.h>
+#  include <map>
 #endif
 
 namespace nix {
