+++ src/libutil/file-system.cc
@@ -27,6 +27,8 @@
 
 #include "nix/strings-inline.hh"
 
+#include "util-config-private.hh"
+
 namespace nix {
 
 namespace fs { using namespace std::filesystem; }
@@ -630,62 +632,6 @@ void replaceSymlink(const fs::path & target, const fs::path & link)
 
         break;
     }
-}
-
-void setWriteTime(
-    const fs::path & path,
-    time_t accessedTime,
-    time_t modificationTime,
-    std::optional<bool> optIsSymlink)
-{
-#ifdef _WIN32
-    // FIXME use `fs::last_write_time`.
-    //
-    // Would be nice to use std::filesystem unconditionally, but
-    // doesn't support access time just modification time.
-    //
-    // System clock vs File clock issues also make that annoying.
-    warn("Changing file times is not yet implemented on Windows, path is %s", path);
-#elif HAVE_UTIMENSAT && HAVE_DECL_AT_SYMLINK_NOFOLLOW
-    struct timespec times[2] = {
-        {
-            .tv_sec = accessedTime,
-            .tv_nsec = 0,
-        },
-        {
-            .tv_sec = modificationTime,
-            .tv_nsec = 0,
-        },
-    };
-    if (utimensat(AT_FDCWD, path.c_str(), times, AT_SYMLINK_NOFOLLOW) == -1)
-        throw SysError("changing modification time of %s (using `utimensat`)", path);
-#else
-    struct timeval times[2] = {
-        {
-            .tv_sec = accessedTime,
-            .tv_usec = 0,
-        },
-        {
-            .tv_sec = modificationTime,
-            .tv_usec = 0,
-        },
-    };
-#if HAVE_LUTIMES
-    if (lutimes(path.c_str(), times) == -1)
-        throw SysError("changing modification time of %s", path);
-#else
-    bool isSymlink = optIsSymlink
-        ? *optIsSymlink
-        : fs::is_symlink(path);
-
-    if (!isSymlink) {
-        if (utimes(path.c_str(), times) == -1)
-            throw SysError("changing modification time of %s (not a symlink)", path);
-    } else {
-        throw Error("Cannot change modification time of symlink %s", path);
-    }
-#endif
-#endif
 }
 
 void setWriteTime(const fs::path & path, const struct stat & st)
