+++ src/nix/self-exe.cc
@@ -1,7 +1,9 @@
 #include "nix/current-process.hh"
 #include "nix/file-system.hh"
 #include "nix/globals.hh"
+
 #include "self-exe.hh"
+#include "cli-config-private.hh"
 
 namespace nix {
